requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pydantic
pdfplumber
python-docx
openai
faiss-cpu
sentence-transformers
